Support C# property syntax
For example... instead of using piece.getType() to determine the type of a
chess piece, you would now use piece.Type instead. The accessor functions
are no longer provided at all since that's simply not "how it works" in C#,
so be prepared to fix your custom client code before you pull this.