Revert "tweaked java templates a bit"
This reverts commit cbf0309bcf3ad1e44aca1e3a1087b32731009ec9.

Rory - I get errors whenever I try to run the java client now.
This commit seems to be to blame. :-( Specifically, somehow we
have it set up so that Main.java can't be in a package... not sure
why, so not sure what to do about it.

I do think we should use Java-style braces in the Java client,
by the way - that's what competitors are expecting! But let's not
commit packages to master unless that gets sorted out.